+++ navigation/Explore/FilteredFeed.tsx
@@ -5,7 +5,6 @@ import { fetchAllBins, fetchBinItemsInfo, BinItemInfo, fetchBinName } from "../.
 import IconWithBackground from "../../components/IconWithBackground";
 import EntypoIcon from "react-native-vector-icons/Entypo";
 import MaterialCommunityIcon from 'react-native-vector-icons/MaterialCommunityIcons';
-import Ionicons from "react-native-vector-icons/Ionicons";
 import { usePostHog } from "posthog-react-native";
 
 interface FilteredFeedProps {
@@ -24,9 +23,6 @@ const FilteredFeed: React.FC<FilteredFeedProps> = ({ navigation }) => {
     useEffect(() => {
       posthog.capture("VIEWED_FILTERED_FEED");
     }, []);
-
-
-
 
 
     useEffect(() => {
