+++ package.json
@@ -22,6 +22,7 @@
     "expo-image-picker": "^14.7.1",
     "expo-status-bar": "~1.11.1",
     "firebase": "^10.8.1",
+    "posthog-react-native": "^3.1.0",
     "react": "18.2.0",
     "react-native": "0.73.4",
     "react-native-async-storage": "^0.0.1",
